Gateway-App/src/test/javascript/spec/app/entities/SmartParking/parking-facility/parking-facility.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { ParkingFacilityService } from 'app/entities/SmartParking/parking-facility/parking-facility.service';
import { IParkingFacility, ParkingFacility } from 'app/shared/model/SmartParking/parking-facility.model';

describe('Service Tests', () => {
  describe('ParkingFacility Service', () => {
    let injector: TestBed;
    let service: ParkingFacilityService;
    let httpMock: HttpTestingController;
    let elemDefault: IParkingFacility;
    let expectedResult: IParkingFacility | IParkingFacility[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ParkingFacilityService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new ParkingFacility(0, 'AAAAAAA', 0, false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ParkingFacility', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new ParkingFacility()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ParkingFacility', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            capacity: 1,
            fullyOccupied: true
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ParkingFacility', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            capacity: 1,
            fullyOccupied: true
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ParkingFacility', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
